ci: move tests function in a common lib
